requirements.txt
streamlit
streamlit-mic-recorder
python-dotenv
gtts
requests
streamlit_oauth
pyjwt